web-scraper-code/requirements.txt
# Core scraping
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data processing
pandas>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Utilities
colorama>=0.4.6
tqdm>=4.66.0

# Visualization
streamlit>=1.30.0
plotly>=5.18.0